+++ src/test/java/com/cobber/fta/TestStrings.java
@@ -47,7 +47,7 @@ public class TestStrings {
 		Assert.assertEquals(result.getMaxValue(), null);
 		Assert.assertEquals(result.getNullCount(), iterations);
 		Assert.assertEquals(result.getBlankCount(), 0);
-		Assert.assertEquals(result.getRegExp(), "[NULL]");
+		Assert.assertEquals(result.getRegExp(), analysis.getRegExp(KnownPatterns.ID.ID_NULL));
 		Assert.assertEquals(result.getConfidence(), 1.0);
 		Assert.assertEquals(result.getType(), PatternInfo.Type.STRING);
 		Assert.assertEquals(result.getTypeQualifier(), "NULL");
@@ -76,7 +76,7 @@ public class TestStrings {
 		Assert.assertEquals(result.getNullCount(), 0);
 		Assert.assertEquals(result.getMatchCount(), 0);
 		Assert.assertEquals(result.getBlankCount(), 3 * iterations + 1);
-		Assert.assertEquals(result.getRegExp(), KnownPatterns.PATTERN_WHITESPACE);
+		Assert.assertEquals(result.getRegExp(), analysis.getRegExp(KnownPatterns.ID.ID_BLANK));
 		Assert.assertEquals(result.getConfidence(), 1.0);
 		Assert.assertEquals(result.getType(), PatternInfo.Type.STRING);
 		Assert.assertEquals(result.getTypeQualifier(), "BLANK");
@@ -107,7 +107,7 @@ public class TestStrings {
 		Assert.assertEquals(result.getNullCount(), 0);
 		Assert.assertEquals(result.getMatchCount(), 0);
 		Assert.assertEquals(result.getBlankCount(), iterations);
-		Assert.assertEquals(result.getRegExp(), KnownPatterns.PATTERN_WHITESPACE);
+		Assert.assertEquals(result.getRegExp(), analysis.getRegExp(KnownPatterns.ID.ID_BLANK));
 		Assert.assertEquals(result.getConfidence(), 1.0);
 		Assert.assertEquals(result.getType(), PatternInfo.Type.STRING);
 		Assert.assertEquals(result.getTypeQualifier(), "BLANK");
@@ -136,7 +136,7 @@ public class TestStrings {
 		Assert.assertEquals(result.getNullCount(), 0);
 		Assert.assertEquals(result.getMatchCount(), 0);
 		Assert.assertEquals(result.getBlankCount(), iterations);
-		Assert.assertEquals(result.getRegExp(), KnownPatterns.PATTERN_WHITESPACE);
+		Assert.assertEquals(result.getRegExp(), analysis.getRegExp(KnownPatterns.ID.ID_BLANK));
 		Assert.assertEquals(result.getConfidence(), 1.0);
 		Assert.assertEquals(result.getType(), PatternInfo.Type.STRING);
 
@@ -167,7 +167,7 @@ public class TestStrings {
 		Assert.assertEquals(result.getMinLength(), 1);
 		Assert.assertEquals(result.getMaxLength(), 9);
 		Assert.assertEquals(result.getBlankCount(), 20);
-		Assert.assertEquals(result.getRegExp(), KnownPatterns.PATTERN_WHITESPACE);
+		Assert.assertEquals(result.getRegExp(), analysis.getRegExp(KnownPatterns.ID.ID_BLANK));
 		Assert.assertEquals(result.getConfidence(), 1.0);
 
 		for (int i = 0; i < inputs.length; i++) {